Refactor : #207 애니메이션 리팩토링 완료
