+++ src/components/index.ts
@@ -3,7 +3,7 @@ export { default as RankingAndRecords } from './RankingAndRecords';
 export { default as Setting } from './Setting';
 export { default as Modal } from './Modal';
 export { default as TodoList } from './TodoList';
-export { default as MainTodo } from './MainTodo';
+export { default as MainTodo } from './MainTodo/MainTodo';
 export { default as AddTodo } from './AddTodo';
 export {
   default as PomodoroTimeSetting,
