+++ src/hooks/useCurrentTodo.ts
@@ -3,6 +3,14 @@ import { TodoEntity } from '../DB/indexedAction';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { timerApi, todosApi } from '../shared/apis';
 import { getDateInFormat } from '../shared/timeUtils';
+import { PomodoroStatus } from '../services/PomodoroService';
+import {
+  IPomodoroActions,
+  IPomodoroData,
+  pomodoroUnit,
+  usePomodoroActions,
+  usePomodoroValue,
+} from './usePomodoro';
 
 interface ITodoFocusedTime {
   id: TodoEntity['id'];
@@ -11,9 +19,17 @@ interface ITodoFocusedTime {
 const TODO_FOCUS_TIME_KEY = 'ExtremeTodoFocusTime';
 type TodoResponseDto = TodoEntity;
 
-const useCurrentTodo = () => {
+const useCurrentTodo = ({
+  value: { settings: pomodoroSettings, status, time },
+  actions,
+}: {
+  value: IPomodoroData;
+  actions: IPomodoroActions;
+}) => {
   const [currentTodo, setCurrentTodo] = useState<TodoResponseDto>();
   const [focusedOnTodo, setFocusedOnTodo] = useState<number>(0);
+  const [canRest, setCanRest] = useState(false);
+  const [shouldFocus, setShouldFocus] = useState(false);
 
   const { data: todos } = useQuery(['todos'], () => todosApi.getList(false), {
     staleTime: 1000 * 60 * 20,
@@ -45,6 +61,70 @@ const useCurrentTodo = () => {
     });
     // TODO : 추후 오프라인일 경우 indexed db에서 현재 투두를 가져와야 할 듯
   }, [todos]);
+
+  useEffect(() => {
+    if (status !== PomodoroStatus.NONE && currentTodo == null) {
+      // actions.stopTimer();
+    } else if (status === PomodoroStatus.NONE && currentTodo != null) {
+      actions.startResting();
+    }
+  }, [currentTodo]);
+
+  useEffect(() => {
+    status !== PomodoroStatus.OVERFOCUSING && checkIfCanRest();
+    checkIfShouldFocus();
+    const ifShouldRest = checkIfShouldRest();
+    (status === PomodoroStatus.FOCUSING ||
+      status === PomodoroStatus.OVERFOCUSING) &&
+      !ifShouldRest &&
+      updateFocus(time === 0 ? 0 : 1000);
+  }, [time]);
+
+  /**
+   * 쉴 수 있는 상황인지(투두에 기록된 duration을 초과했을 때)
+   * @returns boolean
+   */
+  const checkIfCanRest = () => {
+    if (
+      currentTodo?.duration &&
+      focusedOnTodo >=
+        currentTodo?.duration * pomodoroSettings.focusStep * pomodoroUnit
+    ) {
+      setCanRest((prev) => {
+        if (!prev) actions.startResting();
+        return true;
+      });
+      return true;
+    } else {
+      return canRest;
+    }
+  };
+
+  /**
+   * 쉬어야 하는 상황인지(집중 단위시간이 다 되었을 때)
+   * @returns boolean
+   */
+  const checkIfShouldRest = () => {
+    if (
+      status === PomodoroStatus.FOCUSING &&
+      time === pomodoroSettings.focusStep * pomodoroUnit
+    ) {
+      actions.startResting();
+      return true;
+    }
+    return false;
+  };
+
+  const checkIfShouldFocus = () => {
+    if (
+      status === PomodoroStatus.RESTING &&
+      (time ?? 0) >= pomodoroSettings.restStep * pomodoroUnit
+    ) {
+      setShouldFocus(true);
+    } else {
+      setShouldFocus(false);
+    }
+  };
 
   const checkLocalStorageAndGetFocusTime = (todo: TodoEntity) => {
     const storageFocusTime = localStorage.getItem(TODO_FOCUS_TIME_KEY);
@@ -79,6 +159,8 @@ const useCurrentTodo = () => {
       doTodoMutate({ id: currentTodo.id, focusTime: focusedOnTodo });
     getNextTodo();
     setFocusedOnTodo(0);
+    actions.startFocusing();
+    setCanRest(false);
   };
 
   const getNextTodo = (): TodoEntity | undefined => {
@@ -114,6 +196,8 @@ const useCurrentTodo = () => {
       currentTodo,
       focusedOnTodo,
       checkLocalStorageAndGetFocusTime,
+      canRest,
+      shouldFocus,
     }),
     [
       doTodo,
@@ -121,6 +205,8 @@ const useCurrentTodo = () => {
       currentTodo,
       focusedOnTodo,
       checkLocalStorageAndGetFocusTime,
+      canRest,
+      shouldFocus,
     ],
   );
